requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
discord.py==2.5.2
dotenv==0.9.9
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.2.0
propcache==0.3.0
python-dotenv==1.0.1
Werkzeug==3.1.3
yarl==1.18.3